projects/sandbox/train/pyproject.toml
[tool.poetry]
name = "train"
version = "0.1.0"
description = "Generate training data and launch trainer"
authors = ["ethan.marx <ethan.marx@ligo.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
torch = {version = "*", source = "torch"}
"mlpe.data" = {path = "../../../libs/data", develop  = true}
"mlpe.logging" = {path = "../../../libs/logging", develop = true}
"sampling" = {path = "../sampling", develop = true}

"ml4gw" = {path = "../../../ml4gw", develop = true}
typeo = {path = "../../../typeo", develop = true}
lightning = "^2.0"
jsonargparse = {version = "^4.17", extras = ["signatures"]}
ray = "^2.9.1"
pyro-ppl = "^1.8.6"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu116"
secondary = true
default = false

[tool.poetry.scripts]
train = "train.train:main"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
